Move all platform specific source file in subdirs Rename for consistency Adapt build systems
