+++ CMakeLists.txt
@@ -14,15 +14,9 @@ set(CMAKE_CXX_EXTENSIONS OFF)
 
 # Source Files
 add_library(${PROJECT_NAME}
-	src/3ds_clock.cpp
-	src/3ds_clock.h
-	src/3ds_ui.cpp
-	src/3ds_ui.h
 	src/async_handler.cpp
 	src/async_handler.h
 	src/async_op.h
-	src/audio_3ds.cpp
-	src/audio_3ds.h
 	src/audio_al.cpp
 	src/audio_al.h
 	src/audio.cpp
@@ -31,8 +25,6 @@ add_library(${PROJECT_NAME}
 	src/audio_generic.cpp
 	src/audio_generic.h
 	src/audio.h
-	src/audio_psp2.cpp
-	src/audio_psp2.h
 	src/audio_resampler.cpp
 	src/audio_resampler.h
 	src/audio_sdl.cpp
@@ -41,8 +33,6 @@ add_library(${PROJECT_NAME}
 	src/audio_sdl_mixer.h
 	src/audio_secache.cpp
 	src/audio_secache.h
-	src/audio_switch.cpp
-	src/audio_switch.h
 	src/background.cpp
 	src/background.h
 	src/baseui.cpp
@@ -162,10 +152,7 @@ add_library(${PROJECT_NAME}
 	src/image_xyz.cpp
 	src/image_xyz.h
 	src/input_buttons_desktop.cpp
-	src/input_buttons_gekko.cpp
 	src/input_buttons.h
-	src/input_buttons_opendingux.cpp
-	src/input_buttons_psp.cpp
 	src/input.cpp
 	src/input.h
 	src/input_source.cpp
@@ -198,13 +185,8 @@ add_library(${PROJECT_NAME}
 	src/platform_clock.h
 	src/player.cpp
 	src/player.h
-	src/psp2_clock.cpp
-	src/psp2_clock.h
-	src/psp2_ui.cpp
-	src/psp2_ui.h
 	src/rect.cpp
 	src/rect.h
-	src/registry.cpp
 	src/registry.h
 	src/registry_wine.cpp
 	src/rtp.cpp
@@ -282,10 +264,6 @@ add_library(${PROJECT_NAME}
 	src/state.cpp
 	src/state.h
 	src/std_clock.h
-	src/switch_clock.cpp
-	src/switch_clock.h
-	src/switch_ui.cpp
-	src/switch_ui.h
 	src/system.h
 	src/teleport_target.h
 	src/text.cpp
@@ -301,8 +279,6 @@ add_library(${PROJECT_NAME}
 	src/util_macro.h
 	src/utils.cpp
 	src/utils.h
-	src/util_win.cpp
-	src/util_win.h
 	src/version.h
 	src/weather.cpp
 	src/weather.h
@@ -386,6 +362,13 @@ add_library(${PROJECT_NAME}
 	src/window_varlist.h
 )
 
+# These are actually unused when building in CMake
+#	src/sdl_ui.cpp
+#	src/sdl_ui.h
+#	src/platform/wii/wii_input_buttons.cpp
+#	src/platform/opendingux/opendingux_input_buttons.cpp
+#	src/platform/psp/psp_input_buttons.cpp
+
 include(CMakeDependentOption)
 
 # Include directories
@@ -403,8 +386,6 @@ set(PLAYER_BUILD_EXECUTABLE ON)
 
 if(${PLAYER_TARGET_PLATFORM} STREQUAL "SDL2")
 	target_sources(${PROJECT_NAME} PRIVATE
-		src/sdl_ui.cpp
-		src/sdl_ui.h
 		src/sdl2_ui.cpp
 		src/sdl2_ui.h)
 	target_compile_definitions(${PROJECT_NAME} PUBLIC USE_SDL=2)
@@ -413,11 +394,11 @@ if(${PLAYER_TARGET_PLATFORM} STREQUAL "SDL2")
 	target_link_libraries(${PROJECT_NAME} SDL2::SDL2main)
 elseif(${PLAYER_TARGET_PLATFORM} STREQUAL "libretro")
 	target_sources(${PROJECT_NAME} PRIVATE
-		src/audio_libretro.cpp
-		src/audio_libretro.h
-		src/input_buttons_libretro.cpp
-		src/libretro_ui.cpp
-		src/libretro_ui.h)
+		src/platform/libretro/libretro_audio.cpp
+		src/platform/libretro/libretro_audio.h
+		src/platform/libretro/libretro_input_buttons.cpp
+		src/platform/libretro/libretro_ui.cpp
+		src/platform/libretro/libretro_ui.h)
 	target_compile_definitions(${PROJECT_NAME} PUBLIC USE_LIBRETRO=1)
 	set(PLAYER_BUILD_EXECUTABLE OFF)
 	set_target_properties(${PROJECT_NAME} PROPERTIES PREFIX "")
@@ -425,8 +406,42 @@ elseif(${PLAYER_TARGET_PLATFORM} STREQUAL "libretro")
 
 	add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/builds/libretro)
 	target_link_libraries(${PROJECT_NAME} retro_common)
+elseif(${PLAYER_TARGET_PLATFORM} STREQUAL "3ds")
+	target_sources(${PROJECT_NAME} PRIVATE
+		src/platform/3ds/3ds_audio.cpp
+		src/platform/3ds/3ds_audio.h
+		src/platform/3ds/3ds_clock.cpp
+		src/platform/3ds/3ds_clock.h
+		src/platform/3ds/3ds_ui.cpp
+		src/platform/3ds/3ds_ui.h)
+	target_compile_definitions(${PROJECT_NAME} PUBLIC _3DS)
+elseif(${PLAYER_TARGET_PLATFORM} STREQUAL "psvita")
+	target_sources(${PROJECT_NAME} PRIVATE
+		src/platform/psvita/psp2_audio.cpp
+		src/platform/psvita/psp2_audio.h
+		src/platform/psvita/psp2_clock.cpp
+		src/platform/psvita/psp2_clock.h
+		src/platform/psvita/psp2_ui.cpp
+		src/platform/psvita/psp2_ui.h)
+	target_compile_definitions(${PROJECT_NAME} PUBLIC PSP2)
+elseif(${PLAYER_TARGET_PLATFORM} STREQUAL "switch")
+	target_sources(${PROJECT_NAME} PRIVATE
+		src/platform/switch/switch_audio.cpp
+		src/platform/switch/switch_audio.h
+		src/platform/switch/switch_clock.cpp
+		src/platform/switch/switch_clock.h
+		src/platform/switch/switch_ui.cpp
+		src/platform/switch/switch_ui.h)
+	target_compile_definitions(${PROJECT_NAME} PUBLIC __SWITCH__)
 else()
 	message(FATAL_ERROR "Invalid target platform")
+endif()
+
+if(WIN32)
+	target_sources(${PROJECT_NAME} PRIVATE
+		src/registry.cpp
+		src/platform/windows/utils.cpp
+		src/platform/windows/utils.h)
 endif()
 
 if(CMAKE_BUILD_TYPE STREQUAL "Debug")
