+++ src/platform/3ds/3ds_audio.h
@@ -15,8 +15,8 @@
  * along with EasyRPG Player. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef EP_AUDIO_3DS_H
-#define EP_AUDIO_3DS_H
+#ifndef EP_PLATFORM_3DS_AUDIO_H
+#define EP_PLATFORM_3DS_AUDIO_H
 
 #include "audio.h"
