+++ src/platform/psvita/psp2_clock.cpp
