+++ src/platform/psvita/psp2_ui.h
@@ -15,8 +15,8 @@
  * along with EasyRPG Player. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef EP_PSP2_UI_H
-#define EP_PSP2_UI_H
+#ifndef EP_PLATFORM_PSP2_UI_H
+#define EP_PLATFORM_PSP2_UI_H
 
 #ifdef PSP2
