+++ src/platform/switch/switch_clock.cpp
